requirements.txt
pandas
pyarrow
databricks-sql-connector
python-dotenv
databricks
scikit-learn
nltk


